requirements.txt
asgiref==3.8.1
attrs==25.4.0
click==8.3.0
Django==5.1.3
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-resized==1.0.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==24.2
phonenumberslite==8.13.52
pillow==11.1.0
pymodbus==3.8.6
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.37.0
rpds-py==0.27.1
sqlparse==0.5.2
typing_extensions==4.15.0
uritemplate==4.1.1
uvicorn==0.38.0
